Feat. divide create project page to components
