+++ src/components/Form/FormError.jsx
@@ -1,6 +1,6 @@
 import PropTypes from "prop-types";
 
-function FormError({ setShow }) {
+function FormError({ errorMessage, setShow }) {
   const handleShowError = () => {
     setShow("");
   };
@@ -20,9 +20,7 @@ function FormError({ setShow }) {
       >
         <path d="M10 .5a9.5 9.5 0 1 0 9.5 9.5A9.51 9.51 0 0 0 10 .5ZM9.5 4a1.5 1.5 0 1 1 0 3 1.5 1.5 0 0 1 0-3ZM12 15H8a1 1 0 0 1 0-2h1v-3H8a1 1 0 0 1 0-2h2a1 1 0 0 1 1 1v4h1a1 1 0 0 1 0 2Z" />
       </svg>
-      <div className="ms-3 text-sm font-medium">
-        A simple danger alert with an . Give it a click if you like.
-      </div>
+      <div className="ms-3 text-sm font-medium">{errorMessage}</div>
       <button
         type="button"
         className="ms-auto -mx-1.5 -my-1.5 bg-warn-50 text-warn-500 rounded-lg focus:ring-2 focus:ring-warn-400 p-1.5 hover:bg-warn-200 inline-flex items-center justify-center h-8 w-8 dark:bg-gray-800 dark:text-warn-400 dark:hover:bg-gray-700"
@@ -53,6 +51,7 @@ function FormError({ setShow }) {
 
 FormError.propTypes = {
   setShow: PropTypes.func.isRequired,
+  errorMessage: PropTypes.string.isRequired,
 };
 
 export default FormError;
